+++ packages/editor-common/web/src/Components/LinkComponents/LinkPanel.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { mergeStyles, isValidUrl } from 'wix-rich-content-common';
 import Tooltip from 'wix-rich-content-common/libs/Tooltip';
-import Checkbox from '../Checkbox';
+import { Checkbox } from 'wix-rich-content-ui-components';
 import { ErrorIcon } from '../../Icons';
 import styles from '../../../statics/styles/link-panel.scss';
 import { LinkPanelDropdown } from './LinkPanelDropdown';