+++ packages/ricos-editor/web/src/RicosEditor.tsx
@@ -76,11 +76,12 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
 
   componentDidMount() {
     this.updateLocale();
-    const { children } = this.props;
-    const onOpenEditorSuccess =
-      children?.props.helpers?.onOpenEditorSuccess ||
-      this.props._rcProps?.helpers?.onOpenEditorSuccess;
-    onOpenEditorSuccess?.(Version.currentVersion);
+    const { isMobile, toolbarSettings } = this.props;
+    const { useStaticTextToolbar } = toolbarSettings || {};
+    this.getBiCallback('onOpenEditorSuccess')?.(
+      Version.currentVersion,
+      isMobile ? ToolbarType.MOBILE : useStaticTextToolbar ? ToolbarType.STATIC : ToolbarType.INLINE
+    );
     this.props.editorEvents?.subscribe(EditorEvents.RICOS_PUBLISH, this.onPublish);
   }