+++ CHANGELOG.md
@@ -20,12 +20,31 @@
 - `fullscreen`
   - [#2625](https://github.com/wix/ricos/pull/2625) images with download enabled can be downloaded from fullscreen
 
-### :house: Internal
-- `ricos-content`
-  - [#2610](https://github.com/wix/ricos/pull/2610) tiptap converters fixed
-
 </details>
 <hr/>
+
+## 8.44.5 (Jun 20, 2021)
+### :house: Internal
+- `ricos-viewer`
+  - [#2634](https://github.com/wix-incubator/rich-content/pull/2634) `onViewerLoaded` contains `url` field
+- `ricos-editor`
+  - [#2632](https://github.com/wix-incubator/rich-content/pull/2632) `onOpenEditorSuccess` contains `toolbarType` field
+
+## 8.44.4 (Jun 18, 2021)
+### :bug: Bug Fix
+- `editor-common`
+  - [#2629](https://github.com/wix-incubator/rich-content/pull/2629) render wrapper modal (fix's polls insert modal)
+
+## 8.44.3 (Jun 17, 2021)
+### :house: Internal
+- `editor-common`
+  - [#2622](https://github.com/wix-incubator/rich-content/pull/2622) moved textSearchInput/Separator/Dropdown components into ui-components folder
+  - [#2619](https://github.com/wix-incubator/rich-content/pull/2619) moved RadioGroup/InfoIcon/Checkbox components into ui-components folder
+- `ricos-content`
+  - [#2610](https://github.com/wix/ricos/pull/2610) tiptap converters fixed
+  - [#2628](https://github.com/wix/ricos/pull/2628) `extract` & `modify` API exposed via libs folder
+- `editor`
+  - [#2585](https://github.com/wix/ricos/pull/2585) `onToolbarButtonClick` callback hook
 
 ## 8.44.2 (Jun 14, 2021)
 ### :house: Internal